requirements.txt
asgiref==3.7.2
Django==5.0
django-filter==24.2
djangorestframework==3.15.1
et-xmlfile==1.1.0
gunicorn==21.2.0
jdatetime==4.1.1
Markdown==3.6
numpy==1.26.4
openpyxl==3.1.2
packaging==23.1
pandas==2.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-gettext==5.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.1
whitenoise==6.5.0
